app/agents/news_agent.py
from typing import Dict, List, Tuple, Any
import google.generativeai as genai
from langchain.agents import AgentExecutor
from langchain.agents.format_scratchpad import format_to_openai_function_messages
from langchain.agents.output_parsers import OpenAIFunctionsAgentOutputParser
from langchain.prompts import ChatPromptTemplate, MessagesPlaceholder
from langchain.tools import Tool
from langchain.tools.render import format_tool_to_openai_function
from langgraph.graph import StateGraph, END
from pydantic import BaseModel

class NewsAgentState(BaseModel):
    """State for the news processing agent."""
    query: str
    articles: List[Dict[str, Any]] = []
    summaries: List[Dict[str, Any]] = []
    current_step: str = "initial"
    error: str = None

class NewsAgent:
    def __init__(self, gemini_api_key: str):
        """Initialize the news agent with Gemini API."""
        genai.configure(api_key=gemini_api_key)
        self.model = genai.GenerativeModel('gemini-pro')
        
        # Define tools for the agent
        self.tools = [
            Tool(
                name="search_news",
                func=self._search_news,
                description="Search for news articles based on a query"
            ),
            Tool(
                name="summarize_article",
                func=self._summarize_article,
                description="Generate a summary of a news article"
            ),
            Tool(
                name="filter_relevance",
                func=self._filter_relevance,
                description="Filter articles based on relevance to the query"
            )
        ]
        
        # Set up the agent workflow
        self.workflow = self._create_workflow()
    
    def _create_workflow(self) -> StateGraph:
        """Create the LangGraph workflow for news processing."""
        workflow = StateGraph(NewsAgentState)
        
        # Define nodes
        workflow.add_node("search", self._search_news)
        workflow.add_node("summarize", self._summarize_article)
        workflow.add_node("filter", self._filter_relevance)
        
        # Define edges
        workflow.add_edge("search", "filter")
        workflow.add_edge("filter", "summarize")
        workflow.add_edge("summarize", END)
        
        # Set entry point
        workflow.set_entry_point("search")
        
        return workflow
    
    async def _search_news(self, state: NewsAgentState) -> NewsAgentState:
        """Search for news articles based on the query."""
        try:
            # TODO: Implement NewsAPI integration
            # This is a placeholder for the actual implementation
            state.articles = [
                {"title": "Sample Article", "content": "Sample content"}
            ]
            state.current_step = "search_complete"
        except Exception as e:
            state.error = str(e)
        return state
    
    async def _summarize_article(self, state: NewsAgentState) -> NewsAgentState:
        """Generate summaries for the filtered articles."""
        try:
            for article in state.articles:
                prompt = f"""Summarize the following news article in a concise and informative way:
                Title: {article['title']}
                Content: {article['content']}
                """
                
                response = self.model.generate_content(prompt)
                summary = {
                    "title": article["title"],
                    "summary": response.text,
                    "source": article.get("source", "Unknown")
                }
                state.summaries.append(summary)
            
            state.current_step = "summarization_complete"
        except Exception as e:
            state.error = str(e)
        return state
    
    async def _filter_relevance(self, state: NewsAgentState) -> NewsAgentState:
        """Filter articles based on relevance to the query."""
        try:
            filtered_articles = []
            for article in state.articles:
                prompt = f"""Rate the relevance of this article to the query: "{state.query}"
                Article: {article['title']}
                Rate from 0-1 and explain why.
                """
                
                response = self.model.generate_content(prompt)
                # TODO: Implement proper relevance scoring
                if "relevance" in response.text.lower():
                    filtered_articles.append(article)
            
            state.articles = filtered_articles
            state.current_step = "filtering_complete"
        except Exception as e:
            state.error = str(e)
        return state
    
    async def process_query(self, query: str) -> Dict[str, Any]:
        """Process a user query and return relevant news summaries."""
        state = NewsAgentState(query=query)
        result = await self.workflow.arun(state)
        return {
            "summaries": result.summaries,
            "error": result.error
        } 