requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
google-generativeai>=0.3.0
streamlit>=1.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.6.0
newsapi-python>=0.2.7
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 